:recycle: move getHeaderHeight from hooks to utils index
