+++ client/src/components/TutoTwitter/TutoTwitter.tsx
@@ -2,7 +2,8 @@ import React, { FunctionComponent, useEffect, useState } from 'react'
 import ClassNames from 'classnames'
 import './tuto-twitter.scss'
 import Button, { themes as buttonThemes } from '../Button/Button'
-import { useWindowSize, getHeaderHeight } from '../../utils/hooks'
+import { useWindowSize } from '../../utils/hooks'
+import { getHeaderHeight } from '../../utils'
 import { TweenMax } from 'gsap'
 import TutoVideo from '../../assets/video/tuto.mp4'
 import mockup from '../../assets/images/mockup.png'
